+++ CHANGELOG.md
@@ -65,9 +65,9 @@ All notable changes to this project will be documented in this file.
   [#261]).
 - Opt-in message of experimental annotations (PR [#328]).
 - Kotlin compiler's treatment of warnings as errors (issue [#330]).
-- Move the `unaryMinus` **experimental** operations on `AnyInt`, `NonZeroInt`
-  and `NegativeInt` to the `kotools.types.experimental` package with a new
-  signature (issue [#319]).
+- Move the `unaryMinus` **experimental** operations on `AnyInt`, `NonZeroInt`,
+  `PositiveInt` and `NegativeInt` to the `kotools.types.experimental` package
+  with a new signature (issue [#319]).
 
 [4.2.0]: https://github.com/kotools/types/releases/tag/4.2.0
 [#260]: https://github.com/kotools/types/issues/260